moved the component into ui
