+++ package.json
@@ -16,6 +16,7 @@
     "@radix-ui/react-slot": "^1.1.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "embla-carousel-react": "^8.3.1",
     "lucide-react": "^0.426.0",
     "next": "14.2.5",
     "react": "^18",
